refactor: reduce the number of iterations to check that constraints are met
Reduce the number of iterations to check that constraints are met.
When determining whether there is a duplicate among three numbers,
there is no need to calculate the third number because indexOf and lastIndexOf are used.

BREKING CHANGE: Decrease the number of iteration operator.

Before:

for (let i = 0; i < inputData.length; i++){
// ... //
}

After:

for (let i = 0; i < inputData.length - 1; i++){
// ... //
}